kill puppetdb on java out-of-memory so systemd detects failure
This is an attempt to have the puppetdb container restart
automatically via systemd if the jvm runs out of memory.

Before this commit, systemd fails to detect if the java
process crashes due to OOM because systemd is not monitoring
the underlying java process, and the underlying failure
does not bubble up to puppetdb controlling process.

PuppetLabs recommends to bounce puppetdb if jvm runs out of memory,
so killing puppetdb has a similar effect.

After this commit, `/etc/sysconfig/puppetdb` has this param:

    JAVA_ARGS="-Xmx512m -XX:+HeapDumpOnOutOfMemoryError
               -XX:HeapDumpPath=/var/log/puppetdb/puppetdb-oom.hprof
               -Djava.security.egd=file:/dev/urandom
               -Xms192m
               -XX:OnOutOfMemoryError='kill -9 %p'"